bluetooth: support soft volume control for HFP
Based on a patch from Rodrigo Araujo <araujo.rm@gmail.com>;

According to the HFP 1.6 specification, remote volume control isn't
mandatory. If it is, only the speaker volume control is mandatory.

Add flags on the hfp_config struct and check for this case. By default,
we assume software volume control. When the HF sends us its features
during negotiation, if it supports remote volume control then we disable
software VC for the speaker, but we keep assuming software VC for the
mic unless we receive a unsolicited +VGM at some point.

Signed-off-by: Faidon Liambotis <paravoid@debian.org>